Working path for inserting links to studio embeds
